Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/historical.window-expected.txt


PASS registerContentHandler() is removed 
PASS isProtocolHandlerRegistered() is removed 
PASS isContentHandlerRegistered() is removed 
PASS unregisterContentHandler() is removed 
